flask_app/requirements.txt
Flask==1.1.2
google-cloud-ndb==1.8.0
